Area of Square.py
This Python code prompts the user to input a value for the side length of a square. It reads the input as a floating-point number and assigns it to the variable side1. Then, it calculates the area of the square by squaring the value of side1 and assigns it to the variable area. Finally, it prints out the calculated area.